examples/python-client-agent/requirements.txt
requests
python-dotenv
openai
colorama
google-generativeai
a2a
